Final art addition
I really hope I didn't go too excessive on my silly images.
Feel free to remove the ones that suck in a future commit.

Related, I added defaults.cfg as a backup for the default status of
config.cfg, in case there are are any accidental commits of a testing
state of config.cfg.  Folder weights are now in these files.